+++ src/Modulo5.Data/Modulo5DbContext.cs
@@ -14,9 +14,12 @@ public class Modulo5DbContext : DbContext
 
     public DbSet<Perfil> Perfiles => Set<Perfil>();
 
+    public DbSet<Articulo> Articulos => Set<Articulo>();
+
     protected override void OnModelCreating(ModelBuilder modelBuilder)
     {
         modelBuilder.ApplyConfiguration(new PerfilConfiguration());
         modelBuilder.ApplyConfiguration(new UsuarioConfiguration());
+        modelBuilder.ApplyConfiguration(new ArticuloConfiguration());
     }
 }